FIX: Search all function for supplier proposals
In commit https://github.com/Dolibarr/dolibarr/commit/6c8ae0b2a6b10bcf03c6a4b557c29f673dd73f63 the behavior of searching all functionality for supplier proposals via the list.php was changed a bit. It was provided that the supplier_proposal.class.php defines the array of $fields with the flag if a search all is allowed in this field. This short fix adds only the two fields to the list to make the function again working. Without this you cannot search after the reference via the search all functionality.